Now stdin/stdout redirection can be disabled in config.

+++ nsuolymp.py
@@ -39,6 +39,18 @@ class save_cwd():
 		# type: (Any, Any, Any) -> None
 		os.chdir(self.cwd)
 
+# like contextlib.nullcontext in Python 3.7 (better typing?)
+class null_context:
+	def __init__(self, arg = None):
+		# type: (Any) -> None
+		self.arg = arg
+	def __enter__(self):
+		# type: () -> Any
+		return self.arg
+	def __exit__(self, *exc_info):
+		# type: (Any) -> None
+		pass
+
 # returns contents of file by given path (or None if it is not present)
 def read_file_contents(filepath):
 	# type: (Optional[Union[str, bytes]]) -> Optional[bytes]
@@ -701,8 +713,10 @@ def controlled_run_solution(solution, time_limit, memory_limit, interactive, qui
 		return sol_res._replace(verdict = exitcode_verdict)
 	else:
 		proclaim_process_runs([popen_args], [time_limit], [corrected_memory_limit], quiet)
-		with open("input.txt", "rb") as fin:
-			with open("_stdout_", "wb") as fout:
+		cmin = open("input.txt", "rb") if enable_stdinout_redirection else null_context(None)       # type: Any
+		with cmin as fin:
+			cmout = open("_stdout_", "wb") if enable_stdinout_redirection else null_context(None)   # type: Any
+			with cmout as fout:
 				process = psutil.Popen(popen_args, stdin = fin, stdout = fout)
 				res = control_processes_execution([process], [time_limit], [corrected_memory_limit], None, quiet)
 		return res[0]
@@ -842,10 +856,11 @@ def check_solution_on_test(cfg, solution, input_file, gen_output = False):
 	if path.isfile(out_fn):
 		copyfile(out_fn, 'output.txt')
 
-	if getfilesize('_stdout_') == 0:
-		os.remove('_stdout_')
-	if getfilesize('output.txt') <= 0 and getfilesize('_stdout_') > 0:
-		copyfile('_stdout_', 'output.txt')
+	if enable_stdinout_redirection:
+		if getfilesize('_stdout_') == 0:
+			os.remove('_stdout_')
+		if getfilesize('output.txt') <= 0 and getfilesize('_stdout_') > 0:
+			copyfile('_stdout_', 'output.txt')
 	if gen_output:
 		copyfile('output.txt', 'answer.txt')
